+++ gfx/gfx_component.cpp
@@ -45,7 +45,7 @@ namespace gfx {
 	}
 	void GfxComponent::removeEffect(Effect *effect)
 	{
-		for (int i = 0; i < mEffects.size(); i++)
+		for (size_t i = 0; i < mEffects.size(); i++)
 		{
 			if (mEffects[i].get() == effect)
 			{
